chore: Move BaseConfig from pkg/utils to pkg/brimstone
