+++ cmd/brimstone/main.go
@@ -11,11 +11,11 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/davidh-cyberark/brimstone/pkg/brimstone"
 	bs "github.com/davidh-cyberark/brimstone/pkg/brimstone"
 	gg "github.com/davidh-cyberark/brimstone/pkg/gitguardian"
 	hmsl "github.com/davidh-cyberark/brimstone/pkg/hasmysecretleaked"
 	pam "github.com/davidh-cyberark/brimstone/pkg/privilegeaccessmanager"
-	"github.com/davidh-cyberark/brimstone/pkg/utils"
 
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
@@ -47,7 +47,7 @@ type config struct {
 	DbUrl string `env:"DB_URL,required,unset"`
 	Port  uint16 `env:"PORT" envDefault:"9191"`
 
-	utils.BaseConfig
+	brimstone.BaseConfig
 }
 
 func main() {
